float32 and device sync defaults

+++ src/constants.py
@@ -2,7 +2,7 @@ import numpy as np
 
 CANVAS_SIZE = (400, 400)
 ICON_SIZE = (16, 16)
-DTYPE_ACCURACY = np.float64
+DTYPE_ACCURACY = np.float32
 
 COMPONENTS = ["WATER", "AIR", "LAND"]  # Here and not in the model universe because it is required by the GUI and the model
